Update the 'trace' tool to be a better example
The 'trace' tool can be seen as a more complex example of using the
OpenQMC API. That said, the code could do with improvement to show best
practices.

Update the 'trace' tool so that the code is a better example of best
practice when using OpenQMC. Also update the rendered images for the
README and the notebooks.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>