Further improvements to 'trace' to make it cleaner
Some of the scenes could be tweaked to make easier to reason about in
the code. Also the way the domains were interspered in the code meant
they might be lost when a user is skimming the example.

Update the presence and the blur trace scenes so the are easier to
follow and use a new non depth-of-field camera. Also refactor code to
make domain use clearer.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>