Load clang-tools before clang in nix devshell
Clang tools that are included with CC on macOS don't have the needed
wrappers to handle include paths, makeing clang-tidy etc non-functional.

For the devshell (which will be used for CI), load the stdenv without
CC. Then add clang-tools prior to adding back clang, so that it takes
priority. This allows the clang-tidy package to include wrappers.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>