Update the README to provide more help around Nix
The correct command to install OpenQMC using Nix as a package manager
was not listed. Even though this isn't the method most would use with
Nix to add the dependency, it is consistent with other package managers.

Change the 'Nix OS package' section to provide a traditional install
command. Also update the config-module instructions to be clear that an
install from a package manager will provide the default configuration,
and an install from source is required for custom configs.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>