Update the flake inputs to a new version of nixpkgs
For security reasons its advisied to keep flakes inputs updated. After
running the Nix Flake Checker, the nixpkgs version was shown to be older
than 30 days.

Set the hypothesis input to follow the projects top level nixpkgs so
that it can be updated. Then update the common nixpkgs version used for
all derivatives.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>