This is a WIP to use Nix for reproducible CI runs
It uses the actions developed by Determinate Systems to get the Nix
binary and cache. And then runs all commands in a dev shell, that
matches the one used for local development.

It would also be good to add a new build test that uses submodules + a
new example test that installs and depends via Nix.

Current blockers for merging are:

- Nix install not working with self-hosted runners.
- GPU drivers for CUDA are non hermetic / still need a solution.
- Issue with running binaries (linkage to Nix store from previous job).

A concern is the added bandwidth and compute time. We should measure
these and see if the difference in cost is worth the gains.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>