Update toolchain to Nixpkgs 23.11 to resolve #20
Current dependencies are locked against Nixpkg 23.05 which is no longer
the latest version. Updating Nixpkg would assist in keeping the project
in sync with newer LLVM toolchains and other dependencies.

Update Nixpkgs to 23.11, moving the LLVM toolchain and Clang to version
16. Clang Format and Clang Tidy move to 14. This is in sync with default
version available on GitHub ubuntu-22.04 runner images.

Ruff versions avaible in 23.11 allows us to now use this tool to enforce
formatting and linting for Python code. The script files have been
updated.

Also add documentation changes to reflect the changes to the scipt files
and provide guidance on running the Python notebooks.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>