Fix CI checks so that they can fail when needed
Static analysis checks were not failing correctly due to exit on error
not being enabled.

Add this check, and then fix an active error for the static anlysis
where the example CMake projects were being included in the tests. These
were failing as they are not present in the compile_commands.json file,
being seperate CMake projects. This are now filtered out.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>