Add coverage checks on the CI and also locally
There is no support for checking test coverage across the project.
Having this would allow develops to get a clearer insight into the
health of the projects unit testing, as well as guide where testing
is needed.

Add a new CI pipeline that runs on main to produce a coverage badge and
a html report for detailed insight. Also update the nix enviroment with
LLVM and provide a helper script to check coverage locally. Instructions
on how to do so have been added to the CONTRIBUTING.md file.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>