Replace Git Submodules with FetchContent
We use Git Submodules as a backup for dependencies in case they are
not available on the system. However submodules can be a source of
confusion, and replacing them with CMake's FetchContent to download
source directly at setup.

Remove the submodules, align version hashes with those from Nixpkgs
23.11 (apart from oneTBB which has been brought forward to 2021.11.0
to resolve build issues with Windows), and prevent CXX flags from
propigating to dependency builds.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>