Remove mention of Vale from CONTRIBUTION page
We wanted to remove reference to Vale as use of the tool is not common
place enough. This falls more in the camp of individual developers
having their own prefences and using their own tools.

Documentation style will continue to follow the guidelines layed out,
and will assessed through the code review process.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>