Replace the bounded PRNG and update unit tests
The current bounded RNG handler was based on the design from OpenBSD,
which was debiased, however introduced 2 divisions. This could be
improved upon. Testing for this code could also be improved. Some
tests make assumptions about the implementation, while the bounded
RNG statistical test uses floating point operations when it could use
integer arithmetic.

Update the bounded RNG to use the method from Java, that for small
ranges which are the most common, uses close to 1 division on average.
Remove the signed variant as this could be done trivially by the client,
and so is not critical and is out of scope. Update the unit tests so
that they make no assumptions on implementation, and remove any floating
point arithmetic from the the bounded statistical test.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>