Remove the 2^16 index limit from the sampler API
Samplers had a requirement that sample indices could not exceed 2^16 to
aid performance. This was problematic as it adds complexity for a user
and is a potential point of failure. We wanted to remove this limit
while also retaining the performance.

One reason it is impractical for a user to enforce the index limit is
that newDomainSplit() will scale the internal index, compounding for
each call to the function. We can't expect users to track the scaling
ratio and then also correctly handle cases where this becomes invalid.

Remove the index limit constraint on the interface. Instead of changing
the underlying algorithms to support values above 2^16, robustly protect
against high indices by leveraging domain chaining. When a user requests
a sample value above 2^16, branch a new domain based on the quotient,
using the remainder as the internal index value.

Remove the nextDomainIndex() function and restructure the API usage
pattern. This was a required step towards removing the index limit, and
also has the benefit of simplifying the API from a user perspective.

Update the trace tool to use the new API and demonstrate the different
strategies available for a user. This will be useful once creating the
documentation.

Minor changes include: Update some terms like 'samples' to 'sample',
'rnds' to 'rnd', and 'sampleId' to 'index' for better semantics. Also
declare the function pixelDecorrelate() const to match the rest of the
State64Bit type interface.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>