Update the Python config and Jupyter Notebooks
Update the nix development environment so that Python is evaluated with
packages directly, allowing Jupyter to be correctly configured. Also add
a just recipe to launch notebooks once in the development environment.

Update the notebooks and supporting Python code to add new plots for
illustrating the workflows around the recently updated domain splitting
API. These will be used in the upcoming user documentation.

Add new diagrams to support the upcoming documentation. Also include the
original .diagrams files used to author said diagrams, and other images
in the project in case they need re-authoring.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>