Add range based integer remapping to interface
Sampler interfaces do not provide a method to integer remap values
from either the QMC sequence or a PRNG to an integer range. It would
be beneficial to provide dedicated support for such remapping as the
implementation can be nuanced as well as have considerable performance
pitfalls.

Add support on the SamplerInterface to produce both PRNG values and QMC
samples within a ranged interval. Remove redundant functionality from
the pcg.h file and unit tests. Add a new dedicated rnage.h file.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>