Update the README error plots to be more accurate
Error in the README plots is averaged from many runs to make the results more stable and improve clarity on the measurements. However, this could be improved by computing the SE rather than the MAD of the mean estimation. SE is a better fit because CLT shows that the mean is normally distributed.

Error plots have been updated in the notebook and the README. Changes to the just file have also been made to make saving README images easier.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>