Update the Doxyfile configuration for the project
Doxygen needs configuration to provide it information about the project.
This includes the name or project, where to find the files for the
public API, and other options to tailor the output.

Update the default Doxyfile and tailor the configuration to the project.
Create a MAINPAGE.md file for the site's front page. Add graphviz to the
Nix flake to support dot graphs in the documentation.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>