Update documentation on handling sampler state
The documentation is perhaps too technical and does not give clear
enough guidance on sampler state and how it should be managed.

Update and match the documentation in both the README page as-well-as in
the source code for Doxygen output. Give clear guidance on how to pack
samplers when deferring computation to a queue, or passing a sampler to
a function. Remove the sampler implementations from the docs.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>