Add some theoretical explanation around variance
It would be great if there could be a primer to introduce a reader of
the docs to rate of convergence prior to showing them the graphs. Adding
some extra documentation here might allow a reader to interpret the
results and differential between the different samplers more easily, and
make an informed decision.

Add a primer at the beginning of the 'Rate of convergence' section on
the README page. Break this further down into a conceptual description,
some visual graphs, formulas for the asymptotic complexity, and finally
a note on practical expectations.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>