Add new plots and update existing figures
Some plots are currently missing from the documentation. Also, the error
ranges in the existing figures could be more tightly bound.

Add missing plots for the documentation. And update the error bounds as
well as make all colours between plots consistently point to the same
samplers. Add markers for the first M or N plot points for the scatter
plots. Also add RNG to the plots and tools for comparison.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>