Update the formatting used for Doxygen comments
Public header files currently use C style formatting for all Doxygen
comments. This might benefit from a more modern and terse style now
available in Doxygen and C++.

Update to a more modern style of comments that are easier to author, as
well as remove unnecessary 'brief' and 'details' keywords.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>