Add newDomainChain function to sampler interface
A point was made when gathering feedback that it might be benificial to
add a dedicated newDomainChain function to the sampler interface as it
is a distinct strategy outlined in the documentation.

Incorporate this feedback and add a newDomainChain function as shorthand
for chaining to newDomain calls. Also update some of the related docs so
that the use case for this function is clear.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>