Add a newline and whitespace after brief docs
In feedback it was suggested that whitespace between the brief and the
body would make the doxygen style docs clearer to read. Currently, there
is no whitespace.

Add a simple whitespace after the brief across the public API. Also add
a just recipe to build the docs, and make sure the clean recipe removes
the html directory.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>