Add support for running a dev env using Docker
It is common practice in industry to use Docker as a development env,
and it is more likely that a developer will have Docker installed in
comparison to Nix.

Provide an alternative workflow for users without Nix, and that are
more comfortable with Docker. This is simply a wrapper around the Nix
workflow and does not duplicate the maintenance cost. Probably further
work could be done in this area. Might be good to capture feedback
first.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>