Update docker dev env to use docker compose
Current Docker development workflow is not very usable as it copies the
source code rather than mirror it. As a result changes externally are
not represented once the container image is built.

Instead, switch to using Docker Compose to mount the source code as
a volume. This allows the code to stay in sync. Also update the docs.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>