Streamline the development workflow and docs
There are issues with the Docker development environment such as Git
LFS not being active upon startup, and Jupyter Notebooks not port
forwarding. It was also hard coded to either the notebook devShell or
the standard devShell.

Combine both the Nix devShell environments to simplify the workflow
for both 'nix develop' and 'docker compose run develop'. Also make
appropriate changes to the Jupyter Notebook setup so it works from
within the Docker container. Finally, resolve the Git LFS installation
issue and update the docs to make the information more easily
discoverable.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>