Add breadcrumbs to track source of sobol matrices
It would be helpful if the code provided a breadcrumb trail to track the
source of sobol matrices and other similar data that is inlined to the
code. This will then aid and allow review of the project sources.

Add the comments to track forward and backward where the sobol matrices,
rank1 lattice numbers, and optimised blue noise numbers are sourced from
and used. Also provide a navigation sections to the docs.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>