Update Nixpkgs from 23.11 to 24.11
Nixpkgs version is out of date. Updating will keep the compiler tooling
current while also providing a more recent version of ruff for python
static analysis.

Updating the compiler tooling meant that requireing C++14 was not enough
to enforce C++14, because the compiler is allowed to default to a higher
specification. Newer versions of Clang default to C++17. Resolving this
means switching to CMAKE_CXX_STANDARD which is a stronger contract.

Changes to CMake and GTest require updates. CMake curses interface now
needs to use the base preset to allow variables to be updated. While the
newer version of GTest does not support C++14 and so needs to be rolled
back with FetchContent.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>