Add an SVG version of the logo to the images folder
It would be helpful to have the logo in an SVG format for marketing
and other standard use cases on websites. Although there is no original
version of this logo as a vector graphic.

Run the raster image of the logo through a vectorizer and save out the
result as an SVG.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>