Add back per-frame randomization
With the move from STBN to BN, we reduced the time dimension on the BN
tables to a depth of one, and extended their 2D extent. Which is optimal
for a still image. However, this has had the side effect of each frame
in an animation periodically looping on a single 2D table of randomized
pixels.

Over time, this will visually give a static noise look to an animated
sequence. It could also potentially reduce the quality of any downstream
temporal denoising pass.

Add back randomization per-frame by moving the frame argument on the
state constructor from the pixel ID to the initial seed value for the
pattern ID. This works because the encoding format in the state gave
zero bytes to the third access, and so the frame number was not factored
into the encoded pixel ID. By moving it into the pattern ID, it then has
an impact on the resulting sample distribution for a given pattern.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>