Update Doxygen configuration and expose APIs
Module or group pages are hidden under the Topics directory on the API
docs. It would be better if these were more discoverable.

Change the MAINPAGE.md so that the new API groups are easily accessed
from the root of the site. Also update the layout structure and remove
the Topics section, adding direct links to the API groups on the top
level of the navigation.

Also, update the theme of the Doxygen docs to be more modern, using the
popular awesome-doxygen theme.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>