Update logo with latest design
We are in the process of designing a new logo for the project. Desipite
not yet being finalised, it would be good to keep the site up to date
with the latest design.

Update the logo in the images/ directory and link to this from the main
README.md page.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>