Add initial support for MkDocs and CI build automation
We want a main site to support the project, helping those visiting the
site to get started with OpenQMC and find useful information.

Add integration with MkDocs as static site generator. Along with this,
add a new GitHub workflow to automate building and deploying the static
site.

Along with this, deploy the recently updated Doxygen API reference
site - linking to this and removing the inlined API reference from the
README.md file.

Sections and content for the site have been outlined in the mkdocs.yml
file. However, the pages do not currently exist. Next task will be to
automate the generation of those pages from the main README.md file.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>