Add script to split README into pages
For the documentation site we wanted to automate the process of
splitting the README.md file. This would mean that the README and the
site would remain in sync.

Add a new script that takes the README and uses markers in the file
to split the file into individual pages. Also add a new Releases and
Homepage to the site.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>