Update and simplify nix enviroment
The nix development enviroment was on 24.11 and could be updated to the
latest. This would reduce some of the complexity due to resolved issues
in clang tool wrapping, as well as provide access to the latest tooling.

Remove the nixpkgs release target, allowing the project to update to the
latest stable. Targeting a fixed release is not necessary as nix locks
the dependencies and we already target the latest stable.

Along with this, update to the latest Python and Jupyter versions. This
reduces the build time for the Python dependencies, and as a result,
fixes the issue of Docker builds hanging on startup.

Clang compiler has gone from 13 to 21, and with this new clang-tidy
rules were introduced. Some of these rules have been disabled, while
other required code changes for them to pass.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>