Update and simplify Docker setup
The Docker setup required both a Docker and a compose file. The Docker
file was just to install git-lfs so that nix flakes would work. However
this doesn't appear to be an issue anymore, and so we can reference the
nixos image directly from the compose file, removing the Docker file.

This was also a good opertunity to add a Docker volume to store the Nix
cache. This resolves an issue where a user would have to wait for the
cache to fully populate on every start of the container. Now the cache
is persistent, starting a container after the initial run is almost
instantanious. And if any changes to the devshell is made, these will
still be respected.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>