Add contribution docs to site
We don't currently have the content from the CONTRIBUTION.md file
anywhere on the site. It would be good if this appeared under the
Developer section.

Extend the split-readme script so that it takes the CONTRIBUTION.md file
as well and splits this into a page we can reference in the TOC.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>