Start using Nix for consistency on CI
After initial exploration done on '9756110' the conclusion was that
there were a few blocker preventing us from effectively using Nix on the
GitHub CI. Re-visiting this after recent updates, it is now much more
viable to leverage Nix in the CI for consistent testing.

This commit does not deploy Nix for all CI build stages, but it does for
all test jobs. This allows us to drive the clang-tools version from the
locked Nix environment, as opposed to the variable versions used on the
GitHub runners. We then get a bit more compiler coverage on the later
jobs, although this might also get moved over in time.

Updating clang-tools also meant that some tests failed. As a result,
this commit also updates the code to resolve the failing tests.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>