Add ty check for Python type checking
There is a TODO to enable ty from Astral for type checking on the CI.
This wasn't feasible before as the same version was not availble via
the CI runners. Since switching the runners over to Nix, that has now
changed.

Uncomment the lines that enable ty checks on the CI script.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>