Update shell scripts to improve consistency
Shell scripts do not have a consistent shebang. Firstly they don't
reference the environment bash, which could be different depending on
the system. It would be better to enforce that the bash accessed was
that provided by the Nix flake. Secondly the shebang should be the first
line in a file.

Resolve both of these issues, and verify that all scripts continue to
function.

Signed-off-by: Josh Bainbridge <josh.bainbridge@gmail.com>